requirements.txt
Django==4.1
mysqlclient==2.1.1
djangorestframework==3.14.0
markdown==3.3.4
django-filte==22.1
black==19.10b0